+++ mycodo/mycodo_flask/utils/utils_general.py
@@ -1561,11 +1561,60 @@ def get_camera_paths(camera):
     return still_path, tl_path
 
 
+def bytes2human(n, format='%(value).1f %(symbol)s', symbols='customary'):
+    """
+    Convert n bytes into a human readable string based on format.
+    symbols can be either "customary", "customary_ext", "iec" or "iec_ext",
+    see: http://goo.gl/kTQMs
+    
+    Author: Giampaolo Rodola' <g.rodola [AT] gmail [DOT] com>
+    License: MIT
+
+      >>> bytes2human(1.9)
+      '1.0 B'
+      >>> bytes2human(1024)
+      '1.0 K'
+      >>> bytes2human(1048576)
+      '1.0 M'
+      >>> bytes2human(9856, symbols="iec")
+      '9.6 Ki'
+      >>> bytes2human(9856, symbols="iec_ext")
+      '9.6 kibi'
+      >>> bytes2human(10000, "%(value).1f %(symbol)s/sec")
+      '9.8 K/sec'
+      >>> # precision can be adjusted by playing with %f operator
+      >>> bytes2human(10000, format="%(value).5f %(symbol)s")
+      '9.76562 K'
+    """
+    SYMBOLS = {
+        'customary': ('B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'),
+        'customary_ext': ('byte', 'kilo', 'mega', 'giga', 'tera', 'peta', 'exa',
+                          'zetta', 'iotta'),
+        'iec': ('Bi', 'Ki', 'Mi', 'Gi', 'Ti', 'Pi', 'Ei', 'Zi', 'Yi'),
+        'iec_ext': ('byte', 'kibi', 'mebi', 'gibi', 'tebi', 'pebi', 'exbi',
+                    'zebi', 'yobi'),
+    }
+
+    n = int(n)
+    if n < 0:
+        raise ValueError("n < 0")
+    symbols = SYMBOLS[symbols]
+    prefix = {}
+    for i, s in enumerate(symbols[1:]):
+        prefix[s] = 1 << (i+1)*10
+    for symbol in reversed(symbols[1:]):
+        if n >= prefix[symbol]:
+            value = float(n) / prefix[symbol]
+            return format % locals()
+    return format % dict(symbol=symbols[0], value=n)
+
 def get_camera_image_info():
     """ Retrieve information about the latest camera images """
     latest_img_still_ts = {}
+    latest_img_still_size = {}
     latest_img_still = {}
     latest_img_tl_ts = {}
+    latest_img_tl_size = {}
     latest_img_tl = {}
     time_lapse_imgs = {}
 
@@ -1578,6 +1627,9 @@ def get_camera_image_info():
             latest_img_still_ts[each_camera.unique_id] = datetime.fromtimestamp(
                 each_camera.still_last_ts).strftime("%Y-%m-%d %H:%M:%S")
             latest_img_still[each_camera.unique_id] = each_camera.still_last_file
+            file_still_path = os.path.join(still_path, each_camera.still_last_file)
+            if os.path.exists(file_still_path):
+                latest_img_still_size[each_camera.unique_id] = bytes2human(os.path.getsize(file_still_path))
         else:
             latest_img_still[each_camera.unique_id] = None
 
@@ -1592,17 +1644,18 @@ def get_camera_image_info():
         except Exception:
             pass
 
-        except ValueError:
-            latest_time_lapse_img_full_path = None
         if each_camera.timelapse_last_file:
             latest_img_tl_ts[each_camera.unique_id] = datetime.fromtimestamp(
                 each_camera.timelapse_last_ts).strftime("%Y-%m-%d %H:%M:%S")
             latest_img_tl[each_camera.unique_id] = each_camera.timelapse_last_file
+            file_tl_path = os.path.join(tl_path, each_camera.timelapse_last_file)
+            if os.path.exists(file_tl_path):
+                latest_img_tl_size[each_camera.unique_id] = bytes2human(os.path.getsize(file_tl_path))
         else:
             latest_img_tl[each_camera.unique_id] = None
 
-    return (latest_img_still_ts, latest_img_still,
-            latest_img_tl_ts, latest_img_tl, time_lapse_imgs)
+    return (latest_img_still_ts, latest_img_still_size, latest_img_still,
+            latest_img_tl_ts, latest_img_tl_size, latest_img_tl, time_lapse_imgs)
 
 
 def return_dependencies(device_type):