+++ CHANGELOG.md
@@ -13,6 +13,7 @@ If you are upgrading from a version prior to 8.12.10, following the upgrade, you
 ### Features
 
  - Add camera library: libcamera ([#1117](https://github.com/kizniche/mycodo/issues/1117))
+ - Add size of file for last still and timelapse images
  - Add ability to duplicate Inputs
  - Add ability to invert the saved duty cycle for PWM Outputs